+++ nmraspecds/io.py
@@ -1,10 +1,14 @@
 """
 io module of the nmraspecds package.
 """
+import glob
 import os.path
+import re
 
 import aspecd.io
 import linecache
+
+import matplotlib.pyplot as plt
 import nmrglue
 import numpy as np
 
@@ -13,11 +17,27 @@ import nmraspecds.dataset
 import nmraspecds.processing
 
 
+class UnsupportedDataFormatError(Exception):
+    """Exception raised when data format is not supported.
+
+    Attributes
+    ----------
+    message : :class:`str`
+        explanation of the error
+
+    """
+
+    def __init__(self, message=""):
+        super().__init__(message)
+        self.message = message
+
+
 class DatasetImporterFactory(aspecd.io.DatasetImporterFactory):
     """
-    Factory to return the appropiate importer for the dataset.
+    Factory to return the appropriate importer for the dataset.
 
-    The format is currently determined from the type of data.
+    The format is currently determined from the type of data. Abbreviations
+    for the sample names can be used, i.e. ``42`` instead of ``20240816_sa42``.
 
     Raises
     ------
@@ -28,13 +48,27 @@ class DatasetImporterFactory(aspecd.io.DatasetImporterFactory):
     """
 
     def _get_importer(self):
+        # Check characteristic for fitted data.
+        if self.source.endswith(".asc") or os.path.isfile(
+            self.source + ".asc"
+        ):
+            return FittingImporter(source=self.source)
+        # TODO: Maybe improve process of detecting abbreviated sample names.
+        # Source path is completed if the sample name was abbreviated (path
+        # does not exist).
+        if not os.path.exists(self.source):
+            try:
+                path, expno = os.path.split(self.source)
+                basepath, sample = os.path.split(path)
+                name = glob.glob(f"{basepath}/*{str(sample)}*")[0]
+                self.source = os.path.join(name, expno)
+            except IndexError:
+                raise FileNotFoundError
+        # Standard behaviour for Bruker data.
         if os.path.isdir(self.source):
             return BrukerImporter(source=self.source)
         else:
-            if self.source.endswith(".asc") or os.path.isfile(
-                self.source + ".asc"
-            ):
-                return FittingImporter(source=self.source)
+            raise UnsupportedDataFormatError
 
 
 class BrukerImporter(aspecd.io.DatasetImporter):
@@ -79,6 +113,9 @@ class BrukerImporter(aspecd.io.DatasetImporter):
             id: data
             label: My Data
 
+
+    .. versionchanged:: 0.2
+        Type of nucleus is added to axis quantity
     """
 
     def __init__(self, source=None):
@@ -86,7 +123,9 @@ class BrukerImporter(aspecd.io.DatasetImporter):
         self.parameters["type"] = "proc"
         self.parameters["processing_number"] = 1
         self._parameters = None
+        self._raw_parameters = None
         self._data = None
+        self._dimension = None
 
     def _import(self):
         self._check_for_type()
@@ -96,6 +135,7 @@ class BrukerImporter(aspecd.io.DatasetImporter):
         self._create_axes()
         self._get_spectrometer_frequency()
         self._add_nuclei()
+        self._add_axis_metadata()
         self._import_metadata()
 
     def _import_metadata(self):
@@ -114,7 +154,7 @@ class BrukerImporter(aspecd.io.DatasetImporter):
         for key, value in self._parameters["acqus"].items():
             if key.startswith("NUC") and value != "off" and value != 0:
                 nuclei[key] = value
-        for key in sorted(nuclei.keys()):
+        for key in reversed(sorted(nuclei.keys())):
             self._add_nucleus(key)
 
     def _get_spectrometer_frequency(self):
@@ -123,22 +163,60 @@ class BrukerImporter(aspecd.io.DatasetImporter):
         )["procs"]["SF"]
 
     def _create_axes(self):
-        unified_dict = nmrglue.bruker.guess_udic(self._parameters, self._data)
-        unit_converter = nmrglue.bruker.fileiobase.uc_from_udic(unified_dict)
-        self.dataset.data.axes[0].values = unit_converter.ppm_scale()
+        unified_dict = nmrglue.bruker.guess_udic(
+            self._parameters, self._data, strip_fake=False
+        )
+        for dim in np.arange(0, unified_dict["ndim"]):
+            # unified_dict = self._do_referencing_manually(unified_dict, dim)
+            uc = nmrglue.convert.fileiobase.uc_from_udic(unified_dict, dim)
+            ppmsc = uc.ppm_scale()
+            self.dataset.data.axes[dim].values = ppmsc
 
-        self.dataset.data.axes[0].unit = "ppm"
-        self.dataset.data.axes[0].quantity = "chemical shift"
-        self.dataset.data.axes[1].quantity = "intensity"
+    def _do_referencing_manually(self, unified_dict, dim):
+        # TODO: Diese Referenzierung ist noch sehr komisch und funktioniert
+        #  nicht richtig. Eigentlich müsste man an die Daten aus proc2s ran,
+        #  aber die werden nicht eingelesen?
+        # nmrglue appears to stumble with referencing for data produced using newer versions of TopSpin
+        # These two lines set the referencing manually by referring to the processed data dictionary
+        print(
+            "Carrier",
+            unified_dict[0]["car"],
+            unified_dict[1]["car"],
+        )
+        # unified_dict[dim]["obs"] = self._parameters['procs']['SF']
+        unified_dict[dim]["car"] = (
+            self._raw_parameters["acqus"]["SFO1"] - unified_dict[dim]["obs"]
+        ) * 1e6
+        print(
+            "Carrier2",
+            unified_dict[dim]["car"],
+            unified_dict[dim]["car"] / unified_dict[dim]["obs"],
+        )
+        return unified_dict
+
+    def _add_axis_metadata(self):
+        for nr, nucleus in enumerate(self.dataset.metadata.experiment.nuclei):
+            if nr > self._dimension:
+                pass
+            match = re.match(r"(\d+)([A-Za-z]+)", nucleus.type)
+            number = match.group(1)
+            letters = match.group(2)
+            nucleus = f"{{{number}}}{letters}"
+            self.dataset.data.axes[nr].unit = "ppm"
+            self.dataset.data.axes[nr].quantity = f"^{nucleus} chemical shift"
+        self.dataset.data.axes[-1].quantity = "intensity"
 
     def _read_data(self):
         if "pdata" in self.source:
             self._parameters, self._data = nmrglue.bruker.read_pdata(
                 self.source
             )
+            self._raw_parameters, _ = nmrglue.bruker.read_pdata(self.source)
         else:
             self._parameters, self._data = nmrglue.bruker.read(self.source)
-
+        self._dimension = self._data.ndim
+        if self._dimension == 2:
+            self._data = self._data
         self.dataset.data.data = self._data
 
     def _check_for_type(self):
@@ -228,7 +306,7 @@ class FittingImporter(aspecd.io.DatasetImporter):
     """
     Import data from DMFit with experimental and simulated data.
 
-    Data needs to be exported to ascii-format using the " Export spec,
+    Data needs to be exported to ascii-format using the "Export spec,
     model with all lines" command.
 
     The file is composed with three comment lines:
@@ -245,7 +323,7 @@ class FittingImporter(aspecd.io.DatasetImporter):
     The data then follows in the columns. As only the frequency is available as
     metadata, most NMR specific processing steps cannot be performed. The
     data can then be plotted with the special plotter
-    :class:`nmraspecds.plotting.FittingPlotter2D` which povides a color
+    :class:`nmraspecds.plotting.FittingPlotter2D` which provides a color
     scheme that explains the single peaks.
 
 
@@ -258,8 +336,7 @@ class FittingImporter(aspecd.io.DatasetImporter):
     Examples
     --------
     The import of the dataset is performed as usual. Together with a plot in
-    the simplest
-    case, the recipe looks as follows:
+    the simplest case, the recipe looks as follows:
 
     .. code-block:: yaml
 
@@ -268,7 +345,7 @@ class FittingImporter(aspecd.io.DatasetImporter):
             id: fit-data
             label: My Fitted Data
         tasks:
-          - kind: Singleplot
+          - kind: singleplot
             type: FittingPlotter2D
             properties:
               filename: output.pdf
@@ -284,6 +361,7 @@ class FittingImporter(aspecd.io.DatasetImporter):
                 self.source += ".asc"
 
         data = np.loadtxt(self.source, skiprows=3)
+        # data = self._sort_data_maximum(data)
         self.dataset.data.data = data[:, 1:]
         frequency = float(linecache.getline(self.source, 2).strip("##freq "))
         self.dataset.data.axes[0].values = data[:, 0] / frequency
@@ -295,3 +373,14 @@ class FittingImporter(aspecd.io.DatasetImporter):
             frequency
         )
         self.dataset.metadata.experiment.spectrometer_frequency.unit = "MHz"
+
+    def _sort_data_maximum(self, data):
+        axis = data[:, 0]
+        print(data.shape)
+        data_small = data[:, 2:]
+        max_ = np.argmax(data_small, axis=1).astype(int)
+        sorted_max = np.argsort(max_)[::-1]
+        print(sorted_max)
+        data[:, 2:] = data_small[sorted_max]
+        print(data.shape)
+        return data