+++ nmraspecds/plotting.py
@@ -1,7 +1,9 @@
 """
 plotting module of the nmraspecds package.
 """
+import re
 import aspecd.plotting
+import aspecd.annotation
 import numpy as np
 import matplotlib as mpl
 
@@ -56,9 +58,9 @@ class PlotterExtensions:
             )
 
         freq_axis = self.ax.secondary_xaxis(
-            "top", functions=(backward, forward)
+            "top", functions=(forward, backward)
         )
-        freq_axis.set_xlabel(r"$\Delta \nu\ $kHz")
+        freq_axis.set_xlabel(r"$\Delta \nu\ $ / kHz")
 
 
 class SinglePlotter1D(aspecd.plotting.SinglePlotter1D, PlotterExtensions):
@@ -70,7 +72,7 @@ class SinglePlotter1D(aspecd.plotting.SinglePlotter1D, PlotterExtensions):
     ASpecD documentation of the :class:`aspecd.plotting.SinglePlotter1D`
     class for details.
 
-    Furthermore, the class inhertis all functionality from
+    Furthermore, the class inherits all functionality from
     :class:`PlotterExtensions`. See there for additional details.
 
 
@@ -231,7 +233,8 @@ class SinglePlotter2D(aspecd.plotting.SinglePlotter2D, PlotterExtensions):
 
     def __init__(self):
         super().__init__()
-        self.properties.axes.invert = "x"
+        self.properties.axes.invert = ["x", "y"]
+        self.parameters["switch_axes"] = True
 
     def _create_plot(self):
         super()._create_plot()
@@ -560,11 +563,21 @@ class FittingPlotter2D(SinglePlotter2DStacked):
 
     Providing a colormap changes the default colors of the plot.
 
+    The offset of the residues can be defined as absolute value or as
+    percentage of the amplitude of the data. If one or both values were not
+    given, it is tried to set the values as good as possible.
+
 
     Attributes
     ----------
-    attr : :class:`None`
-        Short description
+    parameters['range_residues'] : :class: `list`
+        The range in which the residuals are accounted for in terms of their
+        offset and RMSD calculation.
+
+    parameters['offset_residues'] : :class: `str` or `float`
+        Gives the offset of the residues either as number or in terms of
+        percentage of the residuals' amplitude. The positive number results
+        in a negative offset.
 
     Raises
     ------
@@ -586,25 +599,42 @@ class FittingPlotter2D(SinglePlotter2DStacked):
 
     def __init__(self):
         super().__init__()
+        self.description = "2D Stacked Plot for fitted datasets"
         self.parameters["offset"] = 0
+        self.parameters["offset_residues"] = None
+        self.parameters["range_residues"] = None
+        self.residues = None
+        self.factor = 0.07
+        self.indicator_maxima = False
+        self._annotation = None
+        self._offset = None
+        self._exclude_from_to_dict.append("residues")
 
     def _create_plot(self):
-        # self.properties.colormap = self._create_colormap()
-
+        self._sanitize_data()
         super()._create_plot()
-
         self._change_line_properties()
 
-        ylim_min = self.put_maxima_below_curves()
-        residues = self.dataset.data.data[:, 0] - self.dataset.data.data[:, 1]
-        self.axes.plot(
-            self.dataset.data.axes[0].values,
-            residues + 1.1 * ylim_min,
-            color="steelblue",
-            alpha=0.6,
-        )
-        # self.axes.legend(['experiment', 'simulation', 'single peaks',
-        # 'residual'])
+        self._insert_residues()
+        self._set_maxima()
+        # self.print_rmsd_in_spectrum(residues)
+
+    def _sanitize_data(self):
+        if (
+            self.dataset.data.axes[0].values[-1]
+            > self.dataset.data.axes[0].values[0]
+        ):
+            self.dataset.data.axes[0].values = self.dataset.data.axes[
+                0
+            ].values[::-1]
+            print()
+            self.dataset.data.data = self.dataset.data.data[::-1, :]
+
+    def print_rmsd_in_spectrum(self, residues, x_1, x_2):
+        x_1 = np.where(self.dataset.data.axes[0].values > 100)[0][-1]
+        x_2 = np.where(self.dataset.data.axes[0].values < 20)[0][0]
+        rmsd = np.sqrt(1 / (x_2 - x_1) * np.mean(self.residues[x_1:x_2] ** 2))
+        self.axes.text(90, 20, f"RMSD = {rmsd:.3f}")
 
     def _change_line_properties(self):
         length = self.dataset.data.data.shape[1] - 2
@@ -616,16 +646,7 @@ class FittingPlotter2D(SinglePlotter2DStacked):
             setattr(self.properties.drawings[nr], "linestyle", linestyle_[nr])
             setattr(self.properties.drawings[nr], "alpha", alpha_[nr])
 
-    def put_maxima_below_curves(self):
-        ylim_min, ylim_max = self.axes.get_ylim()
-        delta_ylim = ylim_max + abs(ylim_min)
-        self.axes.set_ylim(3.3 * ylim_min, ylim_max)
-        maxima = self.get_maxima()
-        print(f"Maxima at {maxima} ppm")
-        [self.axes.text(n + 3, 2.5 * ylim_min, f"{n:.0f}") for n in maxima]
-        return ylim_min
-
-    # TODO: Account for stacking dimension
+    # TODO: Account for stacking _dimension
 
     def get_maxima(self):
         max_ppm = []
@@ -644,3 +665,116 @@ class FittingPlotter2D(SinglePlotter2DStacked):
             "Custom cmap", cmaplist, len(cmaplist)
         )
         return cmap
+
+    def _insert_residues(self):
+        self.residues = (
+            self.dataset.data.data[:, 0] - self.dataset.data.data[:, 1]
+        )
+        data = self.dataset.data.data[:, 0]
+        if not self.parameters["range_residues"]:
+            self._get_dataset_ranges_in_figure()
+        elif self.parameters["range_residues"] == [0, 0]:
+            self.parameters["range_residues"] = None
+        if self.parameters["range_residues"]:
+            upper, lower = self.parameters["range_residues"]
+            x_1 = np.where(self.dataset.data.axes[0].values >= upper)[0][-1]
+            x_2 = np.where(self.dataset.data.axes[0].values <= lower)[0][0]
+            self._offset = abs(max(data[x_1:x_2])) + abs(min(data[x_1:x_2]))
+        if self.parameters["offset_residues"]:
+            self._residues_offset = self.parameters["offset_residues"]
+            if (
+                isinstance(self._residues_offset, str)
+                and "%" in self._residues_offset
+            ):
+                percent = (
+                    float(
+                        re.sub(
+                            r"[^0-9.]", "", self.parameters["offset_residues"]
+                        )
+                    )
+                    / 100
+                )
+                if self._offset:
+                    self._offset *= percent
+                else:
+                    self._offset = (
+                        abs(max(self.dataset.data.data[:, 0]))
+                        + abs(min(self.dataset.data.data[:, 0]))
+                    ) * percent
+            else:
+                self._offset = self.parameters["offset_residues"]
+        elif self.parameters["range_residues"]:
+            self._offset *= self.factor
+        else:  # (not self.parameters["range_residues"] and not self.parameters["offset_residues"]):
+            self._offset = abs(max(data)) + abs(min(data)) * self.factor
+        self.parameters["offset_residues"] = self._offset
+        self.axes.plot(
+            self.dataset.data.axes[0].values,
+            self.residues - self._offset,
+            color="steelblue",
+            alpha=0.6,
+        )
+
+    def _get_dataset_ranges_in_figure(self):
+        range_figure = sorted(self.axes.get_xlim())[::-1]
+        range_data = (
+            self.dataset.data.axes[0].values[0],
+            self.dataset.data.axes[0].values[-1],
+        )
+        if range_figure[0] > range_data[0]:
+            upper = range_data[0]
+        else:
+            upper = range_figure[0]
+        if range_figure[-1] < range_data[-1]:
+            lower = range_data[-1]
+        else:
+            lower = range_figure[-1]
+        self.parameters["range_residues"] = [upper, lower]
+
+    def _set_maxima(self):
+        if not self._annotation:
+            self._annotation = aspecd.annotation.Text()
+        maxima = self.get_maxima()
+        print(f"Maxima at {maxima} ppm")
+        self._annotation.parameters["xpositions"] = self._get_x_positions(
+            maxima
+        )
+        self._get_font_y_offset()
+        annotation_offset = -(self._offset * 1.5 + self._font_offset)
+        self._annotation.parameters["ypositions"] = annotation_offset
+        self._annotation.parameters["texts"] = [
+            f"{max_:.0f}" for max_ in maxima
+        ]
+        self.axes.set_ylim(bottom=annotation_offset)
+        for annotation in self.annotations:
+            if (
+                annotation.parameters["texts"]
+                == self._annotation.parameters["texts"]
+            ):
+                return
+        self.annotate(self._annotation)
+
+    def _get_x_positions(self, maxima):
+        raw_positions = [n + 2 for n in maxima]
+        # TODO:
+        # Check the width of the annotation
+        # Compare width with delta x so see if annotation overlaps
+        # Move the first position to higher ppm, the second to lower ppm. (
+        # each half of the overlap)
+        # How to deal with new conflicts? Recursive function?
+        positions = raw_positions
+        return positions
+
+    def _get_font_y_offset(self):
+        self._font_size = mpl.rcParams["font.size"]
+        font_size_inch = self._font_size / 72  # pt to inch in matplotlib
+        font_size_pixels = self.figure.dpi * font_size_inch
+        # if self._residues_offset:
+        #   _offset = 2*self._residues_offset
+
+        ax_pixels = self.ax.transAxes.transform([(0, 0), (0, 1)])
+        ax_extent_pix = ax_pixels[1][1] - ax_pixels[0][1]
+        top_percent = font_size_pixels / ax_extent_pix
+        ylim = self.ax.get_ylim()
+        font_size_data = (ylim[1] - ylim[0]) * top_percent
+        self._font_offset = font_size_data